Update a bunch of files

+++ Content/Projectiles/Common/FlyingMusket.cs
@@ -11,6 +11,7 @@ using Terraria.Audio;
 using Terraria.GameContent;
 using Terraria.ID;
 using Terraria.ModLoader;
+using UniverseOfSwordsMod.Utilities;
 
 namespace UniverseOfSwordsMod.Content.Projectiles.Common
 {
@@ -41,7 +42,7 @@ namespace UniverseOfSwordsMod.Content.Projectiles.Common
         public override void AI()
         {
             NPC npc = Main.npc[(int)TargetIndex];
-            if (!npc.CanBeChasedBy())
+            if (!npc.CanBeChasedBy(this) && !UniverseUtils.IsAValidTarget(npc))
             {
                 Projectile.Kill();
                 return;
@@ -55,8 +56,7 @@ namespace UniverseOfSwordsMod.Content.Projectiles.Common
             {
                 SoundEngine.PlaySound(SoundID.Item11, Projectile.position);
                 Vector2 spawnVel = Vector2.Normalize(npc.Center - Projectile.Center) * 8f;
-                Projectile proj = Projectile.NewProjectileDirect(Projectile.GetSource_FromAI(), Projectile.Center + spawnVel, spawnVel, ProjectileID.Bullet, Projectile.damage / 2, 4f, Projectile.owner);
-                proj.DamageType = DamageClass.Melee;
+                Projectile.NewProjectileDirect(Projectile.GetSource_FromAI(), Projectile.Center + spawnVel, spawnVel, ModContent.ProjectileType<Bullet>(), Projectile.damage / 2, 4f, Projectile.owner);
             }
         }
